added protocol type in the output of timeline. Also the output of the whois and dig will be written to an csv file instead of json.
Will now also check if the compare_file contains valid IP-addresses.

I still need to add a check for .pcap files. Because now it will only check if the file exist and will try to parse it as .pcap file. This may result in an error.

Also need to suppress warnings coming from the whois library.